requirements.txt

asyncpg
nats-py
python-dotenv
